Add manual subscription deletion script
